requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
Faker==4.0.3
mysqlclient==1.4.6
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
